Moved advisors to com.sap.sgs.phosphor.fosstars.advice
Part of #321
