+++ src/main/java/com/sap/sgs/phosphor/fosstars/advice/oss/LgtmScoreAdvisor.java
@@ -1,4 +1,4 @@
-package com.sap.sgs.phosphor.fosstars.model.advice.oss;
+package com.sap.sgs.phosphor.fosstars.advice.oss;
 
 import static com.sap.sgs.phosphor.fosstars.model.feature.oss.OssFeatures.WORST_LGTM_GRADE;
 import static com.sap.sgs.phosphor.fosstars.model.other.Utils.findValue;
