+++ src/main/resources/com/sap/sgs/phosphor/fosstars/advice/oss/OssAdvices.yml
