+++ src/test/java/com/sap/sgs/phosphor/fosstars/advice/AdviceContentStorageTest.java
@@ -1,4 +1,4 @@
-package com.sap.sgs.phosphor.fosstars.model.advice;
+package com.sap.sgs.phosphor.fosstars.advice;
 
 import static com.sap.sgs.phosphor.fosstars.model.feature.example.ExampleFeatures.NUMBER_OF_COMMITS_LAST_MONTH_EXAMPLE;
 import static com.sap.sgs.phosphor.fosstars.model.feature.example.ExampleFeatures.SECURITY_REVIEW_DONE_EXAMPLE;
@@ -8,9 +8,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import com.sap.sgs.phosphor.fosstars.model.RatingRepository;
-import com.sap.sgs.phosphor.fosstars.model.Subject;
 import com.sap.sgs.phosphor.fosstars.model.rating.example.SecurityRatingExample;
-import com.sap.sgs.phosphor.fosstars.model.subject.oss.GitHubProject;
 import java.io.IOException;
 import java.util.List;
 import org.junit.Test;
@@ -19,10 +17,8 @@ public class AdviceContentStorageTest {
 
   @Test
   public void testLoadFromResource() throws IOException {
-    Subject project = new GitHubProject("test", "project");
-
     AdviceContentStorage storage = new AdviceContentStorage(
-        "com/sap/sgs/phosphor/fosstars/model/advice/AdviceContentStorageTest.yml",
+        "com/sap/sgs/phosphor/fosstars/advice/AdviceContentStorageTest.yml",
         RatingRepository.INSTANCE.rating(SecurityRatingExample.class));
 
     List<AdviceContent> advices = storage.advicesFor(SECURITY_REVIEW_DONE_EXAMPLE);
