+++ src/test/java/com/sap/sgs/phosphor/fosstars/advice/oss/OssSecurityAdvisorTest.java
@@ -1,8 +1,8 @@
-package com.sap.sgs.phosphor.fosstars.model.advice.oss;
+package com.sap.sgs.phosphor.fosstars.advice.oss;
 
+import static com.sap.sgs.phosphor.fosstars.model.feature.oss.OssFeatures.USES_LGTM_CHECKS;
 import static com.sap.sgs.phosphor.fosstars.model.feature.oss.OssFeatures.WORST_LGTM_GRADE;
 import static com.sap.sgs.phosphor.fosstars.model.other.Utils.allUnknown;
-import static com.sap.sgs.phosphor.fosstars.model.value.LgtmGrade.A_PLUS;
 import static com.sap.sgs.phosphor.fosstars.model.value.LgtmGrade.B;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -15,11 +15,12 @@ import com.sap.sgs.phosphor.fosstars.model.subject.oss.GitHubProject;
 import com.sap.sgs.phosphor.fosstars.model.value.ValueHashSet;
 import org.junit.Test;
 
-public class LgtmScoreAdvisorTest {
+public class OssSecurityAdvisorTest {
 
   @Test
   public void testBasics() {
-    LgtmScoreAdvisor advisor = new LgtmScoreAdvisor();
+    OssSecurityAdvisor advisor = new OssSecurityAdvisor();
+
     GitHubProject project = new GitHubProject("org", "test");
 
     // no advices if no rating value is set
@@ -32,14 +33,14 @@ public class LgtmScoreAdvisorTest {
     values.update(allUnknown(rating.score().allFeatures()));
     assertTrue(advisor.adviseFor(project).isEmpty());
 
-    // no advices if the LGTM grade is the best
-    values.update(WORST_LGTM_GRADE.value(A_PLUS));
+    // expect an advice if the LGTM checks are not enabled
+    values.update(USES_LGTM_CHECKS.value(false));
     project.set(rating.calculate(values));
-    assertTrue(advisor.adviseFor(project).isEmpty());
+    assertEquals(1, advisor.adviseFor(project).size());
 
     // expect an advice if the LGTM grade is not the best
     values.update(WORST_LGTM_GRADE.value(B));
     project.set(rating.calculate(values));
-    assertEquals(1, advisor.adviseFor(project).size());
+    assertEquals(2, advisor.adviseFor(project).size());
   }
 }
