+++ src/test/resources/com/sap/sgs/phosphor/fosstars/advice/AdviceContentStorageTest.yml
